Update calculator.py
fixed float division error

+++ calculator.py
@@ -12,7 +12,7 @@ def calculator(number1, number2, operator):
     if(operator=="*"):
         return(float(number1*number2))
     if(operator=="/"):
-        return(float(number1/number2))
+        return(number1/number2)
     if(operator=="//"):
         return(number1//number2)
     if(operator=="**"):
